// src/api/generated/core/models/ApplicationDto.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * RTCMS4J Core-service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ApplicationDto
 */
export interface ApplicationDto {
    /**
     * 
     * @type {number}
     * @memberof ApplicationDto
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof ApplicationDto
     */
    namespaceId: number;
    /**
     * 
     * @type {string}
     * @memberof ApplicationDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ApplicationDto
     */
    description: string;
    /**
     * 
     * @type {boolean}
     * @memberof ApplicationDto
     */
    creationByService: boolean;
}

/**
 * Check if a given object implements the ApplicationDto interface.
 */
export function instanceOfApplicationDto(value: object): value is ApplicationDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('namespaceId' in value) || value['namespaceId'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('creationByService' in value) || value['creationByService'] === undefined) return false;
    return true;
}

export function ApplicationDtoFromJSON(json: any): ApplicationDto {
    return ApplicationDtoFromJSONTyped(json, false);
}

export function ApplicationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApplicationDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'namespaceId': json['namespaceId'],
        'name': json['name'],
        'description': json['description'],
        'creationByService': json['creationByService'],
    };
}

export function ApplicationDtoToJSON(json: any): ApplicationDto {
    return ApplicationDtoToJSONTyped(json, false);
}

export function ApplicationDtoToJSONTyped(value?: ApplicationDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'namespaceId': value['namespaceId'],
        'name': value['name'],
        'description': value['description'],
        'creationByService': value['creationByService'],
    };
}